requirements_dev.txt
django-nose==1.1
flake8==1.5
coverage==3.5.3
mox==0.5.3
ipdb
ipdbplugin
